Typecast enum in MT to get rid of warning.

+++ src/MersenneTwister.h
@@ -71,7 +71,7 @@ class MTRand {
 // Data
 public:
 	typedef unsigned long uint32;  // unsigned integer type, at least 32 bits
-	
+
 	enum { N = 624 };       // length of state vector
 	enum { SAVE = N + 1 };  // length of array for save()
 
@@ -232,7 +232,7 @@ inline void MTRand::seed( uint32 *const bigSeed, const uint32 seedLength )
 	initialize(19650218UL);
 	register int i = 1;
 	register uint32 j = 0;
-	register int k = ( N > seedLength ? N : seedLength );
+	register int k = ( (uint32) N > seedLength ? (uint32) N : seedLength );
 	for( ; k; --k )
 	{
 		state[i] =
